+++ src/test/kotlin/com/belle_musica/RegisterUserTest.kt
@@ -2,6 +2,7 @@ package com.belle_musica
 
 import com.belleMusica.schemas.Users
 import database
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -24,16 +25,19 @@ class RegisterUserTest {
         @BeforeAll
         fun restoreDb(): Unit {
             database.delete(Users) { it.id eq it.id }
-
         }
     }
     @BeforeEach
     fun driver(){
-        val chromeOptions = ChromeOptions()
-        chromeOptions.addArguments("--headless")
-        chromeOptions.addArguments("--disable-gpu")
-        driver = ChromeDriver(chromeOptions)
+//        val chromeOptions = ChromeOptions()
+//        chromeOptions.addArguments("--headless")
+//        chromeOptions.addArguments("--disable-gpu")
+        driver = ChromeDriver()
         driver.get("http://localhost:9999/users/new")
+    }
+    @AfterEach
+    fun closeBrowser(){
+        driver.quit()
     }
 
     @Test
@@ -50,10 +54,8 @@ class RegisterUserTest {
             .filter { it.email eq "mail@mail2.com" }
             .firstOrNull()
         assert(ourUser != null)
-//        println(ourUser)
-//        if (ourUser != null) {
-//            assert(ourUser.email == "mail@mail2.com")
-//        }
+        println("AQUIIIII" + ourUser)
+
     }
 
     @Test
@@ -98,21 +100,18 @@ class RegisterUserTest {
         assert(ourUser == 1)
     }
 
-//    @Test
-//    fun `User Not created because of existing username`() {
-//        val email: Unit = driver.findElement(By.id("emailinput")).sendKeys("mail1@mail2.com")
-//        Thread.sleep(2000)
-//        val password: Unit = driver.findElement(By.id("passwordinput")).sendKeys("Password@123")
-//        Thread.sleep(2000)
-//        val username: Unit = driver.findElement(By.id("usernameinput")).sendKeys("Ahmed")
-//        Thread.sleep(2000)
-//        val submit = driver.findElement(By.id("submitbutton")).click()
-//        Thread.sleep(2000)
-//        val ourUser = database.sequenceOf(Users).toList().size
-//        assert(ourUser == 1)
-//
-//    }
+    @Test
+    fun `User Not created because of existing username`() {
+        val email: Unit = driver.findElement(By.id("emailinput")).sendKeys("mail1@mail2.com")
+        Thread.sleep(2000)
+        val password: Unit = driver.findElement(By.id("passwordinput")).sendKeys("Password@123")
+        Thread.sleep(2000)
+        val username: Unit = driver.findElement(By.id("usernameinput")).sendKeys("Ahmed")
+        Thread.sleep(2000)
+        val submit = driver.findElement(By.id("submitbutton")).click()
+        Thread.sleep(2000)
+        val ourUser = database.sequenceOf(Users).toList().size
 
-
-
+        assert(ourUser == 1)
+    }
 }
